requirements.txt
astroid==2.1.0
autopep8==1.4.3
certifi==2018.11.29
chardet==3.0.4
coverage==4.5.2
coveralls==1.5.1
dj-database-url==0.5.0
Django==2.1.5
django-autofixture==0.12.1
django-extensions==2.1.4
docopt==0.6.2
gunicorn==19.9.0
idna==2.8
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pillow==5.4.1
psycopg2-binary==2.7.6.1
pycodestyle==2.4.0
pylint==2.2.2
python-decouple==3.1
pytz==2018.9
requests==2.21.0
six==1.12.0
typed-ast==1.1.1
urllib3==1.24.1
whitenoise==3.3.1
wrapt==1.10.11
